refactor:팩 인덱스 커스텀 훅 분리
